Link `CompilationInfo` only to `ServerMain` and `IndexBuilderMain` (#1497)
Since #734, QLever uses `cmake` magic to generate a file `CompilationInfo.cpp` with information about the commit hash of the current code and the compilation date. This information was then linked to each binary that uses the `Index` class, which includes `ServerMain` and `IndexBuilderMain` as well as most of the test binaries. As a result, re-compilation took rather long even for minor code changes, which is annoying during development.

This is now fixed by only linking this information only to `ServerMain` and `IndexBuilderMain`. All other binaries get some default values, which is fine because they don't need this information anyway.