More tests, i.p. for error thrown in HTTP request handler (#1498)
The code in the `catch` clauses of `HttpServer::session` in `HttpServer.h` was untested so far. It is now cleverly tested by creating a test `HttpServer` with a request handler that throws an expection after its reponse, redirecting the log, and then checking the log messages. Also test the hashing of `DateYearOrDuration` + make some other minor changes.