Add graph info to block metadata (#1482)
If a block contains triples from at most `MAX_NUM_GRAPHS_STORED_IN_BLOCK_METADATA` (default value: 20) different graphs, store the set of these graphs along with the block metadata. Also store the binary information whether the block contains two identical triples belonging to different graphs. Given a set of graphs (which eventually will come from a query via the `FROM` or `GRAPH` clause) then use this metadata to determine whether a block contains relevant triples at all, or whether all the triples in the block are relevant, or whether the block needs to be filtered and or triples need to be deduplicated.

This is an index-breaking change (it changes the binary format of some of the index files).