Fix URL-decoding bug from previous commit  (#1517)
For URL-decoding, now use (the well documented) `boost::urls::parse_origin_form` instead of the (internal and poorly documented) `boost::urls::parse_query` because the latter turned out to not unescape `+` to space.

TODO for a separate PR: When the URL path is malformed, currently a Boost-internal error message is returned, like `leftover [boost.url.grammar:4]`. Instead an exception should be thrown and caught and a proper error message displayed.