Fold points directly into the `Id` (#1506)
Points (specified by WKT literals with type point) are now stored directly in the `ValueId`, with 30 bits of precision for each of the two coordinates.  This makes working on point geometries (e.g. spatial queries that only work on centroids) orders of magnitude faster, as we save the retrieving of the points from disk as well as parsing them.