Initial implementation of `SpatialJoin` (#1248)
Implement a an Operation `SpatialJoin` that takes two inputs, one variable per input, and a maximal distance `d`. This operation joins each row pair [left, right] where the spatial distance between the value of the two variables is less or equal than `d`.
It can currently be called by adding a triple using the special predicate `<max-distance-in-meters:x>`, where `x` is a positive integer that is used as the maximal distance. For example:

```
SELECT * WHERE {
 ?a <is-a> <tram-stop> .
 ?b <is-a> <restaurnat> .
 ?a <max-distance-in-meters:100> ?b
}
```
Current limitations (will be fixed in the future)
* Only supports points (no linestrings, polygons, etc).
* The implementation is a simple quadratic loop over the cartesian product of both inputs.
* The spatial join should not be called via a special predicate, but via `FILTER(geo:distance(?a, ?b) < 100)`.