Fix unescaped literals in SERVICE VALUES-clause (#1530)
Fixes the bug where literals including unescaped quotes (or other characters non conform with the specification of Rdf-Literals, e.g. `"Abraham "Bram" Wiertz"@en`)  were passed unescaped to the `SERVICE` endpoint (using the `VALUES` clause optimization) causing the `SERVICE` clause to fail with a parsing error of the `SERVICE` query.
Also fixes the behavior of undefined values and blank nodes for this optimization.