Greedy query planning for large connected components (#1442)
So far, the query planning has taken very long when the query graph has a large connected component. With this change, the query planner counts the number of connected subgraphs for each connected component. If that number exceeds a certain budget, the query planning for the component is done using so-called Greedy Operator Ordering (GOO). For a connected component with n nodes, GOO considers a quadratic number of plans, whereas the optimal dynamic programming potentially considers an exponential number of plans. The budget can be controlled via the runtime parameter `query-planning-budget` (default: 1500).