Add BlankNode support for SERVICE (#1504)
With this commit, QLever supports to add new blank nodes during the evaluation of a query. This function is used to support blank nodes in the result of a `SERVICE` request. These blank nodes are distinct from all blank nodes in the index, and also from all blank nodes from other SERVICE request, eve if they came from the same server. This behavior is correct accordin to the SPARQL 1.1 federated query standard.