Add generic chunked fill and copy algorithms (#1564)
These are replacements for `std::ranges::fill` and `std::ranges::copy` that perform their respective loops in chunks of a dedicated size. After each chunk, a user-defined callback is invoked, which can be used e.g. for cancellation checks.
The helpers are used within the `UNION` and `BIND` operations to add more cancellation checks there.