Add Path Search Feature to Qlever (#1335)
This commit adds a feature that computes all paths between pairs of `(source, target)` on a given graph. The source, target, and graph can be configured. The syntax to trigger this search is using a SERVICE request with a special IRI. The details of this API are described in the file `docs/path_search.md`.
The current implementation only supports the computation of all paths, which can be exponentially larger than the number of edges in the graph. The implemented infrastructure makes it simpler to implement additional features for this service in the future, such as:
* Only return an arbitrary single path between given start and end nodes.
* Only return the shortest path
* Only return the longest path (same complexity as "all paths" , but less memory requirements).
* Return all edges that lie on an path.

Additionally it can be extended by efficiently supporting the LIMIT/OFFSET clauses and lazy evaluation to efficiently support dealing with very large results in the future.