Make parallel parsing the default again for a single input stream (#1585)
With the introduction of mutiple input streams (through #1537), the default regarding "parallel-parsing" became "false". That was unfortunate because most existing configurations (in particular, all the Qleverfiles from https://github.com/ad-freiburg/qlever-control) do not set the "parallel-parsing" option explicitly. Without parallel parsing, indexing is much slower. This felt like a regression for many users, for example, see #1563.

This is now fixed as follows: When there is a single input stream, and "parallel-parsing" is neither specified in the `settings.json` file (which is deprecated) nor on the command line for `IndexBuilderMain` (new since #1537), then the default is "true". A warning is shown that this is deprecated. The QLever CLI and Qleverfiles from https://github.com/ad-freiburg/qlever-control will be adapted to avoid this deprecated behavior.

Co-authored-by: Hannah Bast <bast@informatik.uni-freiburg.de>