Only compute the sibling of a `SERVICE` once (#1556)
Previously, the sibling of a `SERVICE` (see #1341 for details) was computed twice. This was not a problem for small siblings result, because they were cached, making the second computation actually a cheap cache read. It was however a problem if the result was too large for the cache. It then was computed twice, although only one of the computations was actually used, because the SERVICE discarded the result because it was too large.
This commit fixes this behavior. The sibling is now only computed once, and only gets explicitly shared with the `SERVICE` if it is sufficiently small.