Add internal function `ql:isGeoPoint` (#1565)
The function returns true if and only if the argument is a `ValueId` with datatype `Datatype::GeoPoint`, that is, it represents a WKT point and the value is encoded directly in the `Id`. This is useful for https://github.com/ad-freiburg/qlever-petrimaps, which does not have to pre-fetch and store the literals for such points.

It is important to note that there are WKT points for which this function evaluates to false. For example, points where the coordinates are out of range, are not encoded in the `Id` but have an entry in the vocabulary. Due to this QLever-specific semantics, we make this a QLever-internal function.

This is our first QLever-internal function. So far we only had QLever-internal predicates (like `ql:has-predicate` or `ql:langtag`) and Qlever-internal entities (like `ql:@en`). Used that occasion to make the names of several of the QLever-internal constants in the code more meaningful and consistent.