Prefiltering on `CompressedBlockMetadata` using binary search (#1503)
Implement efficient relational filters (`<=>`) on a sorted vector of `CompressedBlockMetadata`. These can be used to efficiently
filter out blocks, for which we can guarantee that they don't contain a single element for which the relational filter returns true.
In the future, these prefilters will be applied when such a filter is applied directly on an `IndexScan` .
For example (simplified, in reality blocks contain triples), if a block contains elements in the range `[3, 7]` (which can be deduced from the metadata) and the filter condition is ` <= 2 `  we can  discard the whole block by only looking on its metadata. On the other hand, if (considering the same block) the filter condition is `== 4` then the block can not be filtered out, because we cannot know in advance, if the full block contains the element `4`.