Make `CartesianProductJoin` lazy. (#1609)
This allows the `CartesianProductJoin` operation to produce lazy results as well as to consume a single lazy result itself.
Currently only the last child of the Cartesian product can be consumed lazily, but this can in principle be changed in the future. In particular, the most memory-saving way would be to lazily consume the largest child that supports lazy evaluation.