Merge directly without vector in Filter class (#1633)
This commit affects the case when the input to a FILTER is lazy, but the result is fully materialized. Now the local vocabs of the lazy input blocks are directly merged into the local vocab of the result. Previously they were first stored in a vector, which had a higher RAM footprint for the case of many redundant and typically empty local vocabs in the input.