Fix quadratic runtime of `FILTER` with lazy input and materialized result (#1634)
After #1478, a `FILTER` with lazy input and materialized result had quadratic runtime because of repeated calls to `reserve`. This is now fixed by calling `resize` instead. In the special case of a single block, we simply move it and avoid the copy. On the side, refactor `IdTable::insertAtEnd` to insert column by column and not row by row.