Allow operations to not store their result in the cache (#1665)
Each operation now has a `bool` that determines whether the results can be stored in the cache or not (whether it is actually stored depends on other circumstances, like the available cache size). That `bool` does not have to be fixed when the operation is created, but can be changed.

For example, this is useful for index scans that only return a subset of their full result (because of another constraining operation, like a join or a filter).