Add option `--parser-buffer-size` for `IndexBuilderMain` (#1698)
So far, the buffer size of the parser was hard-coded in `src/index/ConstantsIndexBuilding.h` to `10 MiB`. This was too little for datasets like OSM, which have some very large triples (due to WKT literals of complex geometries). There is now an option `--parser-buffer-size` with which the default buffer size can be overridden.

While at it, externalize all IRIs and literals (except for the Qlever-internal ones) by default. It has been the default in all our Qleverfiles for quite some time now.